+++ posthog_flutter/lib/posthog_flutter_web.dart
@@ -430,8 +430,8 @@ class PosthogFlutterWeb extends PosthogFlutterPlatformInterface {
       return await handleWebMethodCall(
         MethodCall('captureException', {'properties': normalizedData}),
       );
-    } on Exception catch (exception) {
-      printIfDebug('Exception in captureException: $exception');
+    } catch (error) {
+      printIfDebug('Exception in captureException: $error');
     }
   }