+++ posthog_flutter/lib/src/posthog_flutter_version.dart
@@ -1,3 +1,3 @@
 // This file is auto-updated by scripts/bump-version.sh
-const postHogFlutterVersion = '5.36.1';
+const postHogFlutterVersion = '5.36.3';
 const postHogFlutterSdkName = 'posthog-flutter';